fix(SafeAppsTx): handle origin prop correctly and make sure to cal onSubmit
